requirements.txt
pyspark==3.4.1
boto3==1.28.62
psycopg2-binary==2.9.7
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.2
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
great-expectations==0.17.19